extract topic data into a utility function
