+++ app/topics/page.tsx
@@ -2,18 +2,10 @@ import "css/prism.css"
 import "katex/dist/katex.css"
 
 import { MDXLayoutRenderer } from "pliny/mdx-components"
-import {
-    sortPosts,
-    coreContent,
-    allCoreContent
-} from "pliny/utils/contentlayer"
-import { allTopics, allAuthors } from "contentlayer/generated"
-import type { Authors, Topic } from "contentlayer/generated"
-import { notFound } from "next/navigation"
 import TopicLayout from "@/components/bitcoin-topics/layouts/TopicLayout"
 import { components } from "@/components/bitcoin-topics/markdown-ui/MDXComponents"
 import { genPageMetadata } from "../seo"
-import { getAuthorDetails } from "@/utils/content-utils"
+import { getTopicData } from "@/utils/content-utils"
 
 export const metadata = genPageMetadata({
     title: "Bitcoin Topics",
@@ -48,26 +40,8 @@ export const metadata = genPageMetadata({
 const slug = "welcome"
 
 export default async function Page() {
-    // Filter out drafts in production
-    const sortedCoreContents = allCoreContent(sortPosts(allTopics))
-    const postIndex = sortedCoreContents.findIndex((p) => p.slug === slug)
-    if (postIndex === -1) {
-        return notFound()
-    }
-
-    const prev = sortedCoreContents[postIndex + 1]
-    const next = sortedCoreContents[postIndex - 1]
-    const post = allTopics.find((p) => p.slug === slug) as Topic
-    const authorList = post?.authors || ["default"]
-    const authorDetails = getAuthorDetails(authorList)
-    const mainContent = coreContent(post)
-    const jsonLd = post.structuredData
-    jsonLd["author"] = authorDetails.map((author) => {
-        return {
-            "@type": "Person",
-            name: author.name
-        }
-    })
+    const { prev, next, post, authorDetails, mainContent, jsonLd } =
+        getTopicData(slug)
 
     return (
         <>
